Fix Couples, Add W-marked Man/Woman Couples
I just confirmed that there's indeed skin colored couples on Emoji 12, so there it is, fixed the couple position and added the w-marked version of it.
I never explained what w does, but they work like this, w is basically x but after the x-conversion is done, the z-conversion replaces the neon person with more skin colored ones.