It's 2020!
Updated the copyright year, because it's 2020!
If you are wondering if BCF is dead, it was.
But worry not! Active progress will come back soon.
Once again, drastic changes will be done, making the workflow different than before.
The compiler won't return until an alternative is found, Mutant Standard's forc looks good but they depreciated support for SVGinOT exporting, which is bad for us, i will check on how OpenMoji is dealing with them.